Add files via upload
RSA (Rivest–Shamir–Adleman) is a widely used public-key encryption algorithm for secure data transmission over the internet,The encryption process involves converting the plaintext message into a series of numerical values, which are then encrypted using the recipient's public key. The resulting ciphertext can only be decrypted by the recipient using their private key.
The RSA algorithm is based on the fact that it is difficult to factorize large numbers into their prime factors. This means that the security of RSA depends on the difficulty of factoring large numbers.
Here also in the implementation we are required to enter 2 prime numbers reason being The two prime numbers, p and q, play a crucial role in the rsa encryption algorithm. They are used to generate a public and private key pair that is used for encrypting and decrypting messages.
In the rsa encryption algorithm, the public key is a pair of two integers (n, h), where n is the product of the two primes p and q, and h is a random integer between 1 and n-1 that is coprime with (p-1) * (q-1).
The private key is a pair of two integers (p, q), which are the two primes used to generate the public key. The private key is kept secret and is used for decrypting the encrypted message.